feat: added shooter subsystem

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/shooter/ShooterSubsystem.java
@@ -44,8 +44,15 @@ public class ShooterSubsystem {
         hw.shooter.setVelocity(0);
     }
 
-    public void setMaxRPM(double maxRPM){
+    public void setMaxRPM(int maxRPM){
         targetRPM = maxRPM;
     }
 
 }
+
+
+
+
+
+
+
